Added a sell/buy all/half feature
Made a function get_amount inside market_screen that reads a String from the user and goes on with the sell/buy if it's "all", "half" or an integer. Made a function is_int inside gyinput since is_num checks only if the String is a number and it doesn't check if it's an integer.